refactor(lox-analysis): stage the eager analyses in legacy and promote the pipeline to the canonical names
