+++ crates/lox-analysis/src/imaging/aoi.rs
@@ -220,7 +220,13 @@ impl Aoi {
 
 #[cfg(test)]
 mod tests {
+    // Explicit imports shadow the glob, keeping these tests on the eager
+    // implementation until they are ported (see `crate::legacy`).
     use super::*;
+    #[allow(unused_imports)]
+    use crate::legacy::imaging::{
+        AccessAnalysis, AccessResults, OpticalAccessAnalysis, SarAccessAnalysis,
+    };
     use geo::{Distance as GeoDistance, Geodesic, LineString, Polygon, point};
     use lox_bodies::{Earth, MeanRadius};
