+++ crates/lox-analysis/src/imaging/optical.rs
@@ -81,6 +81,13 @@ impl AccessPayload for OpticalPayload {
 
 #[cfg(test)]
 mod tests {
+    // Explicit imports shadow the glob, keeping these tests on the eager
+    // implementation until they are ported (see `crate::legacy`).
+    #[allow(unused_imports)]
+    use crate::legacy::imaging::{
+        AccessAnalysis, AccessResults, OpticalAccessAnalysis, SarAccessAnalysis,
+    };
+
     use super::*;
 
     use geo::{LineString, Polygon};
